+++ examples/simple.rs
@@ -77,13 +77,13 @@ fn main() {
     pushrod_window.window.set_max_fps(60);
     pushrod_window.window.set_ups(60);
 
-    let mut base_widget = PushrodBaseWidget::new();
+    let mut base_widget = BaseWidget::new();
     base_widget.set_origin(Point { x: 100, y: 100 });
     base_widget.set_size(pushrod::core::point::Size { w: 200, h: 200 });
     base_widget.set_color([0.5, 0.5, 0.5, 1.0]);
     pushrod_window.add_widget(Box::new(base_widget));
 
-    let mut box_widget = PushrodBoxWidget::new();
+    let mut box_widget = BoxWidget::new();
     box_widget.set_origin(Point { x: 350, y: 100 });
     box_widget.set_size(pushrod::core::point::Size { w: 200, h: 200 });
     box_widget.set_color([0.0, 1.0, 0.0, 1.0]);