+++ src/widget/widget.rs
@@ -62,7 +62,7 @@ pub enum PushrodWidgetConfig {
     BorderWidth { thickness: u8 },
 }
 
-/// Implementable trait that is used by every `PushrodWidget`.  These are the public methods,
+/// Implementable trait that is used by every `Widget`.  These are the public methods,
 /// and a function _may_ override them.
 ///
 /// You _must_ implement the following methods:
@@ -74,17 +74,17 @@ pub enum PushrodWidgetConfig {
 ///
 /// You _should_ override `draw`, but you are not required to.
 ///
-/// If you want a blank base widget, refer to the `PushrodBaseWidget`, which will create a
+/// If you want a blank base widget, refer to the `BaseWidget`, which will create a
 /// base widget that paints the contents of its bounds with whatever color has been
 /// specified with `set_color`.
-pub trait PushrodWidget {
+pub trait Widget {
     /// Retrieves the configuration HashMap that stores the configuration list of settings
     /// for this widget.
     ///
     /// To implement this, the following code could be used in your object's structure:
     ///
     /// ```
-    /// # use pushrod::widget::widget::PushrodWidget;
+    /// # use pushrod::widget::widget::Widget;
     /// # use pushrod::widget::widget::PushrodWidgetConfig;
     /// # use std::collections::HashMap;
     /// # use std::cell::RefCell;
@@ -104,7 +104,7 @@ pub trait PushrodWidget {
     /// And in the overridden function for get_config in your implementation, use:
     ///
     /// ```
-    /// # use pushrod::widget::widget::PushrodWidget;
+    /// # use pushrod::widget::widget::Widget;
     /// # use pushrod::widget::widget::PushrodWidgetConfig;
     /// # use std::collections::HashMap;
     /// # use std::cell::RefCell;
@@ -113,7 +113,7 @@ pub trait PushrodWidget {
     ///   config: RefCell<HashMap<u8, PushrodWidgetConfig>>
     /// }
     ///
-    /// impl PushrodWidget for MyWidget {
+    /// impl Widget for MyWidget {
     ///
     ///   fn get_config(&mut self) -> &RefCell<HashMap<u8, PushrodWidgetConfig>> {
     ///     &self.config
@@ -238,15 +238,15 @@ pub trait PushrodWidget {
     }
 }
 
-/// This is the `PushrodBaseWidget`, which contains a top-level widget for display.  It does
+/// This is the `BaseWidget`, which contains a top-level widget for display.  It does
 /// not contain any special logic other than being a base for a display layer.
-pub struct PushrodBaseWidget {
+pub struct BaseWidget {
     config: RefCell<HashMap<u8, PushrodWidgetConfig>>,
 }
 
 /// Implementation of the constructor for the `PushrodBaseWidget`.  Creates a new base widget
 /// that can be positioned anywhere on the screen.
-impl PushrodBaseWidget {
+impl BaseWidget {
     pub fn new() -> Self {
         Self {
             config: RefCell::new(HashMap::new()),
@@ -254,7 +254,7 @@ impl PushrodBaseWidget {
     }
 }
 
-/// Implementation of the `PushrodBaseWidget` object with the `PushrodWidget` traits implemented.
+/// Implementation of the `BaseWidget` object with the `Widget` traits implemented.
 /// This function only implements `get_config`, and samples of `mouse_entered`, `mouse_exited`,
 /// and `mouse_scrolled`, which currently trigger messages to the screen.
 ///
@@ -273,7 +273,7 @@ impl PushrodBaseWidget {
 /// #           .unwrap_or_else(|error| panic!("Failed to build PistonWindow: {}", error)),
 /// #   );
 /// #
-///    let mut base_widget = PushrodBaseWidget::new();
+///    let mut base_widget = BaseWidget::new();
 ///
 ///    base_widget.set_origin(Point { x: 100, y: 100 });
 ///    base_widget.set_size(pushrod::core::point::Size { w: 200, h: 200 });
@@ -283,7 +283,7 @@ impl PushrodBaseWidget {
 ///    pushrod_window.add_widget(Box::new(base_widget));
 /// # }
 /// ```
-impl PushrodWidget for PushrodBaseWidget {
+impl Widget for BaseWidget {
     fn get_config(&mut self) -> &RefCell<HashMap<u8, PushrodWidgetConfig>> {
         &self.config
     }